functionality related to format negotiation on verbose errors moved to a separate file
